requirement.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.19.0
matplotlib>=3.3.0
albumentations>=1.0.0
pandas>=1.1.0